// frontend/src/app/api/models/device.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Bios } from '../models/bios';
import { Cpu } from '../models/cpu';
import { Drive } from '../models/drive';
import { Kernel } from '../models/kernel';
import { Mainboard } from '../models/mainboard';
import { OperatingSystem } from '../models/operating-system';
import { Type } from '../models/type';
export interface Device {
  bios?: Bios;
  cpu?: Cpu;
  drives?: Array<Drive>;
  eol?: Date;
  hostname?: string;
  id: string;
  kernel?: Kernel;
  mainboard?: Mainboard;
  manufacturer: string;
  model: string;
  notes?: string;
  os?: OperatingSystem;
  ram?: number;
  storage?: number;
  type: Type;
}